[Channel] Allow channel collector to be enabled without debug mode, extract fake channel support
