+++ src/Sylius/Bundle/ChannelBundle/Collector/ChannelCollector.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Bundle\ChannelBundle\Context\FakeChannel;
+namespace Sylius\Bundle\ChannelBundle\Collector;
 
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Channel\Context\ChannelNotFoundException;
@@ -22,13 +22,14 @@ use Symfony\Component\HttpKernel\DataCollector\DataCollector;
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
-final class FakeChannelCollector extends DataCollector
+final class ChannelCollector extends DataCollector
 {
     /**
      * @param ChannelRepositoryInterface $channelRepository
      * @param ChannelContextInterface $channelContext
+     * @param bool $channelChangeSupport
      */
-    public function __construct(ChannelRepositoryInterface $channelRepository, ChannelContextInterface $channelContext)
+    public function __construct(ChannelRepositoryInterface $channelRepository, ChannelContextInterface $channelContext, $channelChangeSupport = false)
     {
         $this->data['channels'] = $channelRepository->findAll();
 
@@ -37,6 +38,8 @@ final class FakeChannelCollector extends DataCollector
         } catch (ChannelNotFoundException $exception) {
             $this->data['current_channel'] = null;
         }
+
+        $this->data['channel_change_support'] = $channelChangeSupport;
     }
 
     /**
@@ -56,6 +59,14 @@ final class FakeChannelCollector extends DataCollector
     }
 
     /**
+     * @return bool
+     */
+    public function isChannelChangeSupported()
+    {
+        return $this->data['channel_change_support'];
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function collect(Request $request, Response $response, \Exception $exception = null)
@@ -67,6 +78,6 @@ final class FakeChannelCollector extends DataCollector
      */
     public function getName()
     {
-        return 'sylius.context.channel.fake_channel.collector';
+        return 'sylius.collector.channel';
     }
 }
