+++ src/Sylius/Bundle/ChannelBundle/DependencyInjection/Configuration.php
@@ -43,7 +43,7 @@ class Configuration implements ConfigurationInterface
         $rootNode
             ->children()
                 ->scalarNode('driver')->defaultValue(SyliusResourceBundle::DRIVER_DOCTRINE_ORM)->end()
-                ->booleanNode('fake_channel_support')->defaultFalse()->cannotBeEmpty()->end()
+                ->booleanNode('debug')->defaultValue('%kernel.debug%')->cannotBeEmpty()->end()
             ->end()
         ;
