+++ src/Sylius/Bundle/ChannelBundle/Resources/config/debug.xml
@@ -16,7 +16,6 @@
         <parameter key="sylius.context.channel.fake_channel.persister.class">Sylius\Bundle\ChannelBundle\Context\FakeChannel\FakeChannelPersister</parameter>
         <parameter key="sylius.context.channel.fake_channel.code_provider.class">Sylius\Bundle\ChannelBundle\Context\FakeChannel\FakeChannelCodeProvider</parameter>
         <parameter key="sylius.context.channel.fake_channel.context.class">Sylius\Bundle\ChannelBundle\Context\FakeChannel\FakeChannelContext</parameter>
-        <parameter key="sylius.context.channel.fake_channel.collector.class">Sylius\Bundle\ChannelBundle\Context\FakeChannel\FakeChannelCollector</parameter>
     </parameters>
 
     <services>
@@ -32,12 +31,6 @@
             <argument type="service" id="sylius.repository.channel" />
             <argument type="service" id="request_stack" />
             <tag name="sylius.context.channel" priority="128" />
-        </service>
-
-        <service id="sylius.context.channel.fake_channel.collector" class="%sylius.context.channel.fake_channel.collector.class%" public="false">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.context.channel" />
-            <tag name="data_collector" template="SyliusChannelBundle:Collector:channel.html.twig" id="sylius.context.channel.fake_channel.collector" />
         </service>
     </services>
 </container>
