+++ src/Sylius/Bundle/ChannelBundle/Resources/config/services.xml
@@ -20,6 +20,8 @@
 
         <parameter key="sylius.context.channel.request_based.resolver.composite.class">Sylius\Component\Channel\Context\RequestBased\CompositeRequestResolver</parameter>
         <parameter key="sylius.context.channel.request_based.resolver.hostname_based.class">Sylius\Component\Channel\Context\RequestBased\HostnameBasedRequestResolver</parameter>
+
+        <parameter key="sylius.collector.channel.class">Sylius\Bundle\ChannelBundle\Collector\ChannelCollector</parameter>
     </parameters>
 
     <services>
@@ -41,6 +43,13 @@
         <service id="sylius.context.channel.request_based.resolver.hostname_based" class="%sylius.context.channel.request_based.resolver.hostname_based.class%" public="false">
             <argument type="service" id="sylius.repository.channel" />
             <tag name="sylius.context.channel.request_based.resolver" />
+        </service>
+
+        <service id="sylius.collector.channel" class="%sylius.collector.channel.class%" public="false">
+            <argument type="service" id="sylius.repository.channel" />
+            <argument type="service" id="sylius.context.channel" />
+            <argument>false</argument>
+            <tag name="data_collector" template="SyliusChannelBundle:Collector:channel.html.twig" id="sylius.collector.channel" />
         </service>
     </services>
 </container>
